+++ R/preprocess.R
@@ -1,0 +1,3 @@
+preprocess <- function(X, center = FALSE, scale = FALSE, impute = FALSE) {
+    .Call(C_preprocess, X, center, scale, impute, FALSE)
+}