+++ src/init.c
@@ -5,10 +5,12 @@
 
 extern SEXP summarize(SEXP);
 extern SEXP rayOLS(SEXP, SEXP);
+extern SEXP preprocess(SEXP, SEXP, SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef callMethods[] = {
     {"summarize", (DL_FUNC) &summarize, 1},
     {"rayOLS", (DL_FUNC) &rayOLS, 2},
+    {"preprocess", (DL_FUNC) &preprocess, 5},
     {NULL, NULL, 0}
 };