+++ NEWS.md
@@ -1,5 +1,6 @@
 # BGData 2.1.0.9000
 
+- Add `preprocess()` function for fast centering, scaling, and imputation.
 - Follow [Bioconductor S4 practices][1].
   - If you have used `new()` to create `BGData` instances, please use the
     `BGData()` constructor function instead.